+++ integration/airflow/openlineage/airflow/extractors/manager.py
@@ -6,6 +6,7 @@ from openlineage.airflow.utils import get_job_name
 
 
 class ExtractorManager:
+    """Class abstracting management of custom extractors."""
     def __init__(self):
         self.extractors = {}
         self.extractor_mapper = Extractors()