+++ integration/airflow/tests/integration/airflow/dags/event_order.py
@@ -2,8 +2,8 @@ import datetime
 import os
 import uuid
 
-from airflow.operators.bash import BashOperator
-from airflow.operators.python import PythonOperator
+from airflow.operators.bash_operator import BashOperator
+from airflow.operators.python_operator import PythonOperator
 from airflow.utils.dates import days_ago
 from openlineage.client import set_producer, OpenLineageClient
 from openlineage.client.run import RunEvent, Run, Job, RunState