+++ integration/spark/src/main/common/java/io/openlineage/spark/agent/lifecycle/plan/InsertIntoDataSourceVisitor.java
@@ -1,8 +1,11 @@
 package io.openlineage.spark.agent.lifecycle.plan;
 
+import static io.openlineage.spark.agent.facets.TableStateChangeFacet.StateChange.OVERWRITE;
+
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableMap.Builder;
 import io.openlineage.client.OpenLineage;
+import io.openlineage.spark.agent.facets.TableStateChangeFacet;
 import io.openlineage.spark.agent.util.PlanUtils;
 import io.openlineage.spark.api.OpenLineageContext;
 import io.openlineage.spark.api.QueryPlanVisitor;
@@ -24,10 +27,10 @@ public class InsertIntoDataSourceVisitor
 
   @Override
   public List<OpenLineage.OutputDataset> apply(LogicalPlan x) {
+    InsertIntoDataSourceCommand command = (InsertIntoDataSourceCommand) x;
 
     return PlanUtils.applyFirst(
-            context.getOutputDatasetQueryPlanVisitors(),
-            ((InsertIntoDataSourceCommand) x).logicalRelation())
+            context.getOutputDatasetQueryPlanVisitors(), command.logicalRelation())
         .stream()
         // constructed datasets don't include the output stats, so add that facet here
         .peek(
@@ -36,6 +39,9 @@ public class InsertIntoDataSourceVisitor
                   ImmutableMap.<String, OpenLineage.DatasetFacet>builder();
               if (ds.getFacets().getAdditionalProperties() != null) {
                 facetsMap.putAll(ds.getFacets().getAdditionalProperties());
+              }
+              if (command.overwrite()) {
+                facetsMap.put("tableStateChange", new TableStateChangeFacet(OVERWRITE));
               }
               ds.getFacets().getAdditionalProperties().putAll(facetsMap.build());
             })