+++ integration/spark/src/test/resources/spark_scripts/spark_v2_append.py
@@ -1,7 +1,9 @@
 import os
-from pyspark.sql import SparkSession
+import time
 
-os.makedirs("/tmp/v2", exist_ok=True)
+os.makedirs("/tmp/append_data", exist_ok=True)
+
+from pyspark.sql import SparkSession
 
 spark = SparkSession.builder \
     .master("local") \
@@ -10,17 +12,19 @@ spark = SparkSession.builder \
     .config("spark.sql.catalog.spark_catalog.type", "hive") \
     .config("spark.sql.catalog.local", "org.apache.iceberg.spark.SparkCatalog") \
     .config("spark.sql.catalog.local.type", "hadoop") \
-    .config("spark.sql.catalog.local.warehouse", "/tmp/v2") \
+    .config("spark.sql.catalog.local.warehouse", "/tmp/append_data") \
     .config("spark.sql.extensions", "org.apache.iceberg.spark.extensions.IcebergSparkSessionExtensions") \
     .getOrCreate()
 spark.sparkContext.setLogLevel('info')
 
-df = spark.createDataFrame([
-    {'a': 1, 'b': 2},
-    {'a': 3, 'b': 4}
-])
+df = spark.createDataFrame([{'a': 1}, {'a': 2}])
 df.createOrReplaceTempView('temp')
+
 spark.sql("CREATE TABLE local.db.source1 USING iceberg AS SELECT * FROM temp")
 spark.sql("CREATE TABLE local.db.source2 USING iceberg AS SELECT * FROM temp")
 
-spark.sql("CREATE TABLE local.db.target USING iceberg AS (SELECT * FROM local.db.source1 UNION SELECT * FROM local.db.source2)")
+spark.sql("CREATE TABLE local.db.append_table (a long) USING iceberg")
+spark.sql("INSERT INTO local.db.append_table (SELECT * FROM local.db.source1 UNION SELECT * FROM local.db.source2);")
+
+
+